requirements.txt

## install dependencies with "pip install -r requirements.txt"

matplotlib
numexpr
numpy
pandas
pillow
pyyaml
scipy

